+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 * Gaucho now runs as a single-instance app
 * Overhaul of menu and context menu options
     * Suite-related actions moved to context menu
-    * "Import suite" is not part of settings
+    * "Import suite" is now part of settings
 * Global Env variables can be set in settings
+    * These will be exported/imported along with the suites in "Export Tasks"
 * Variable env form validation
 * Adds cut-copy-paste items to context menu
 * Settings shortcuts menu is now collapsed by default